backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Graph Database
neo4j==5.15.0
py2neo==2021.2.4

# Cache & Queue
redis==5.0.1
celery==5.3.4

# AI/ML - Core
openai==1.6.1
anthropic==0.8.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-community==0.0.13
sentence-transformers==2.2.2
litellm==1.17.9

# AI/ML - Multi-Agent
autogen==0.2.0
crewai==0.1.0

# Vector Database
pinecone-client==2.2.4
weaviate-client==3.25.3

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
websockets==12.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Data Pipeline Dependencies
arxiv==1.4.8
scholarly==1.7.11
PyGithub==1.59.1
scrapy==2.11.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
pymupdf==1.23.8

# NLP & Text Processing
spacy==3.7.2
transformers==4.35.2
nltk==3.8.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Async utilities
asyncio-throttle==1.0.2
tenacity==8.2.3